requirements.txt
crewai
langgraph
langchain
llama-index
python-dotenv
requests
openpyxl
rich
tqdm
pydantic
tenacity
pyyaml
pytest
